FIX : Création d'un sprint par défaut pour tous les projets existants

+++ script/migration_stories.php
@@ -36,6 +36,7 @@ foreach($TData as $fk_project => $stories) {
 	$TStorieLabel = explode(',', $stories);
 
 	if(empty($TStorieLabel)) {
+		// Dans le cas où le projet n'utilisait pas l'extrafields "stories", on insère pour ce projet un sprint par défaut
 		$sql = 'INSERT INTO '.MAIN_DB_PREFIX.'projet_storie(fk_projet, storie_order, label)';
 		$sql .= ' VALUES('.$fk_project.', 1, "Sprint 1")';
 
@@ -43,6 +44,7 @@ foreach($TData as $fk_project => $stories) {
 		if(! $resql) $error++;
 	}
 	else {
+		// Sinon, on lui réaffecte ceux qu'il utilisait
 		foreach($TStorieLabel as $k => $storie_label) {
 			$sql = 'INSERT INTO '.MAIN_DB_PREFIX.'projet_storie(fk_projet, storie_order, label)';
 			$sql .= ' VALUES('.$fk_project.', '.($k+1).', "'.ltrim($storie_label).'")';
@@ -55,7 +57,10 @@ foreach($TData as $fk_project => $stories) {
 
 if(empty($error)) {
 	$extrafields=new ExtraFields($db);
-	$extrafields->delete('stories', 'projet');
+	$extralabels = $extrafields->fetch_name_optionals_label('projet');
+	if(! empty($extralabels['stories'])) {
+		$extrafields->delete('stories', 'projet');
+	}
 }
 
 function getData() {
@@ -68,16 +73,18 @@ function getData() {
 		return array();
 	}
 
-	$sql = 'SELECT fk_object, stories';
-	$sql .= ' FROM '.MAIN_DB_PREFIX.'projet_extrafields';
-//	$sql .= ' WHERE stories IS NOT NULL';
+	// Sélectionne tous les projets existants qui n'ont pas de sprint de créé dans la table "projet_storie"
+	$sql = 'SELECT p.rowid, pe.stories';
+	$sql .= ' FROM '.MAIN_DB_PREFIX.'projet AS p';
+	$sql .= ' LEFT JOIN '.MAIN_DB_PREFIX.'projet_extrafields AS pe ON pe.fk_object=p.rowid';
+	$sql .= ' WHERE p.rowid NOT IN (SELECT fk_projet FROM '.MAIN_DB_PREFIX.'projet_storie)';
 
 	$resql = $db->query($sql);
 
 	$TData = array();
 	if($resql) {
 		while ($obj = $db->fetch_object($resql)) {
-			$TData[$obj->fk_object] = $obj->stories;
+			$TData[$obj->rowid] = $obj->stories;
 		}
 	}
